Поменял у Артикулов тип данных кодов. В Меню отображаю кол-во Задач и Рекламаций. Переписал запрос на получение комментов на ЗИР. Пофиксил ЗИР так же как и на стороне 1С.

+++ app/src/main/java/ua/com/merchik/merchik/data/Database/Room/ArticleSDB.java
@@ -21,7 +21,7 @@ public class ArticleSDB {
     @SerializedName("vendor_code")
     @Expose
     @ColumnInfo(name = "vendor_code")
-    public Integer vendorCode;
+    public Long vendorCode;
 
     @SerializedName("tovar_id")
     @Expose
